Fixed gender issue
Issue was that the temp variable I was using for getting a random number was an int so when I divided it by 100, it always rounded down to 0 and the logic in determining gender makes it so that if the result is less than the genderRatio (in Bulbasaur's case 0.875), it would be male.  So 0 would always be less than 0.875 so the pokemon was always male.  Fixed by casting the temp to a double at the moment of dividing it by 100 to compare it to the genderRatio.

+++ src/OwnedPokemon.java
@@ -39,7 +39,7 @@ public class OwnedPokemon {
 			gender = "genderless";
 		else {
 			temp = new Random().nextInt(100) + 1;
-			if (temp/100 <= pokemon.getGenderRatio())
+			if ((((double) temp)/100) <= pokemon.getGenderRatio())
 				gender = "male";
 			else
 				gender = "female";
